requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
cffi==1.12.3
Click==7.0
cryptography==2.7
decorator==4.4.0
Django==2.2.5
djangorestframework==3.10.3
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pytz==2019.2
shell==1.0.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
